third_party/blink/web_tests/fragmentation/single-line-cells.html

<!DOCTYPE html>
<style>
table {
    border-collapse: collapse;
}
td {
    background-color: #ddd;
    border: 1px solid black;
}
tr {
    break-inside: avoid;
}
.td { color: blue; }
</style>
<p>crbug.com/99124: Table rows shouldn't straddle page boundaries.</p>
<div style="columns:3; line-height: 20px; column-fill: auto; gap:0; width:450px; height:190px; background-color: yellow;">
    <table>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td class="td" data-total-x="151" data-total-y="1">Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td class="td" data-total-x="301" data-total-y="1">Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
    </table>
</div>
<div id="console"></div>
<script src="../resources/check-layout.js"></script>
<script>
    checkLayout(".td", document.getElementById("console"));
</script>
